Add Consul Watch configuration to Docker Compose
This configuration monitors the checks of a given service or those in a
specific state. The service `gentry` compiles the application and
creates a binary executable in the `/go/bin` directory. The `consul`
service invokes the binary executable in the `watches` configuration
defined in the `CONSUL_LOCAL_CONFIG` environment variable.

Closes #3